apps/ai-gateway/pyproject.toml
[project]
name = "ai-gateway"
version = "0.0.1"
description = "Python AI Gateway for LLM completions and embeddings"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.9",
  "structlog>=24.1",
  "httpx>=0.27",
  "litellm>=1.40.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
  "opentelemetry-exporter-otlp-proto-http>=1.26.0",
  "opentelemetry-instrumentation-logging>=0.47b0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "mypy>=1.12",
  "ruff>=0.6",
  "black>=24.8",
  "pip-licenses>=4.5.1"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
warn_return_any = true

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

